+++ library/graphics/Shape.cc
@@ -136,7 +136,7 @@ namespace gf {
         const std::size_t offset = vertices.size();
         vertices.insert(vertices.end(), other.vertices.begin(), other.vertices.end());
         std::transform(other.indices.begin(), other.indices.end(), std::back_inserter(indices), [offset](uint16_t index) {
-          return index + offset;
+          return static_cast<uint16_t>(index + offset);
         });
       }
     };
